+ Added Radius type for SearchRadius objects * renamed resource dir "range_search" => "locations" * extracted searchradii into * Refactored external data (e.g. tags,locations) into resource dir
