+++ app/src/models/radius.ts
@@ -1,0 +1,6 @@
+export interface Radius {
+  text: string;
+  value: string;
+}
+
+export default Radius;
