+++ app/src/resources/locations/index.js
