+++ app/src/resources/locations/latloncitysize_enrichment.py
