+++ app/src/resources/radii/index.ts
@@ -1,0 +1,10 @@
+import Radius from '@/models/radius';
+
+import radiiData from './radii.json';
+const radii: Radius[] = radiiData as Radius[];
+
+export {
+  radii
+};
+
+export default radii;
