+++ app/src/resources/radii/radii.json
@@ -1,0 +1,22 @@
+[
+  {
+    "text": "Überall",
+    "value": ""
+  },
+  {
+    "text": "5 km",
+    "value": "5km"
+  },
+  {
+    "text": "10 km",
+    "value": "10km"
+  },
+  {
+    "text": "25 km",
+    "value": "25km"
+  },
+  {
+    "text": "50 km",
+    "value": "50km"
+  }
+]
