+++ app/src/resources/tags/index.js
