+++ app/tslint.json
@@ -5,7 +5,8 @@
   ],
   "linterOptions": {
     "exclude": [
-      "node_modules/**"
+      "node_modules/**",
+      "src/resources/**/*.json"
     ]
   },
   "rules": {
